Replace use of TArray with std::vector in FIWadManager
This is part of an ongoing project to replace the use of custom
containers types with standard library types. In this case, I've
replaced the use of the custom TArray resizable array class with the
standard std::vector type. The semantics of the two types are very
similar, with only a little adjustment required to get the standard
library equivelant to work.