Rename FString::GetChars() to FString::c_str()
This is the first move in removing the FString class. We'll start by
trying to get the existing FString class to mimic std::string as closely
as possible, so at the end we can switch out FString for std::string
without any issues.

We'll start with the simplest move, renaming GetChars to c_str.