Replace FString::IsNotEmpty() with !FString::empty()
This one is another straightforward replacement. The std::string class
contains an empty() method, but nothing for the inverse. Since I'm
trying to conform the FString API to the std::string API for a later
swap, we need to remve the use of FString::IsNotEmpty(). This change
swaps out IsNotEmpty for the an inverted empty() call.